Add Kubernetes generators
